backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-core>=1.2.7",
    "langchain-deepseek>=1.0.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0",
    "uvicorn>=0.40.0",
]
